amplify/functions/transcribeAudio/requirements.txt
faster-whisper==1.1.1
pyannote.audio==3.3.2
torch==2.3.1
torchaudio==2.3.1
boto3==1.34.0
httpx==0.27.0
pydub==0.25.1
imageio-ffmpeg==0.5.1
requests==2.32.3
huggingface_hub==0.23.0
filelock==3.14.0
packaging==24.0
tqdm==4.66.4
PyYAML==6.0.1
typing_extensions==4.12.2
sympy==1.12
networkx==3.3
